requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1.1
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-filter==22.1
django-rest-auth==0.9.5
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-mimeparse==1.6.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
RestAuthCommon==0.7.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
